third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/script-tag-null-char-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cp%3E%00%00%00%00%00%00%00%3Cscript%3Ealert(%27%00%27)%3C/script%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
